% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtRTD2.R
\name{AtRTD2.example}
\alias{AtRTD2.example}
\title{Save AtRTD2 data into csv files}
\usage{
AtRTD2.example(dir = ".")
}
\arguments{
\item{dir}{directory to save data. If the directory does not exist, a new folder will be created with the provided name.}
}
\value{
Three csv tables of corresponding data in AtRTD2.
}
\description{
Save AtRTD2 data into csv files
}
\examples{
AtRTD2.example(dir='data')

}

