% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtRTD2.R
\docType{data}
\name{AtRTD2}
\alias{AtRTD2}
\title{Example datasets of AtRTD2.}
\format{An object of data list}
\usage{
data(AtRTD2)
}
\description{
The data sources include subset of TPM time-series transcript expression data with 26 time points, 3 biological replicates and 3
technical replicates (234 sample points in total) from the AtRTD2 paper (Zhang, et al., 2016), a gene-isoform mapping table of 300
genes (first column) and 766 transcripts (second column) and a subset of 100 isoforms.
}
\examples{
names(AtRTD2)
##expression data
AtRTD2$data.exp[1:10,1:5]
##mapping data
AtRTD2$mapping[1:10,]
##subset of isoform names
AtRTD2$sub.isoforms[1:10]
}
\references{
Zhang, R., et al. AtRTD2: A Reference Transcript Dataset for accurate quantification of alternative splicing and expression changes
in Arabidopsis thaliana RNA-seq data. bioRxiv 2016.
}
\keyword{datasets}

