% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch.density.R
\name{switch.density}
\alias{switch.density}
\title{Density plot}
\usage{
switch.density(x, time.points, make.plotly = T, plot.type = "density",
  show.line = T, title = "Density of switch points", ...)
}
\arguments{
\item{x}{a numeric vector to plot th density.}

\item{time.points}{a numeric vector of the time points of time-series, e.g. 1,2,3,...}

\item{make.plotly}{logical, to plot \code{\link{plotly}} format figures (TRUE)
or general plot (FALSE)?. See details in \code{\link{ggplotly}} in \code{\link{plotly}} R pacakge.}

\item{plot.type}{the plot types. Options are "Density_line" for density line plot, "Density_bar" for density bar plot and "Frequency_bar" for frequency bar plot.}

\item{...}{additional parameters pass to \code{\link{density}}}
}
\value{
density plot in ggplot2 format or \code{\link{plotly}} format if \code{make.plotly=T}
}
\description{
Density plot
}
