% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.spline.R
\name{ts.spline}
\alias{ts.spline}
\title{Fit a time-series with spline curve}
\usage{
ts.spline(x, t.start, t.end, nrep, df = 5, ...)
}
\arguments{
\item{x}{a vector of time-series expression}

\item{t.start, t.end}{start and end time points of the time-series data. The time step is assumed to be 1.}

\item{nrep}{number of replicates.}

\item{df}{degree of freedom used in \code{\link{ns}} in \code{\link{splines}} package.}

\item{...}{additional arguments passed to \code{\link{predict}}.}
}
\value{
predictions results returned from \code{\link{predict}}.
}
\description{
Fit a time-series with spline curve
}
\seealso{
\code{\link{lm}}, \code{\link{ns}}, \code{\link{predict}}
}
