% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSIS.app.R
\name{TSIS.app}
\alias{TSIS.app}
\title{Isoform switch analysis and visualization with Shiny App}
\usage{
TSIS.app(data.size.max = 100)
}
\arguments{
\item{data.size.max}{maximum size limit for unload files in Shiny. Default is 100 (MB).}
}
\value{
the Shiny App.
}
\description{
Isoform switch analysis and visualization with Shiny App
}
\examples{
TSIS.app()

}
\seealso{
\code{\link{shiny}}
}

