% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSIS.data.example.R
\name{TSIS.data.example}
\alias{TSIS.data.example}
\title{Save TSIS.data data into csv files}
\usage{
TSIS.data.example(dir = ".")
}
\arguments{
\item{dir}{directory to save data. If the directory does not exist, a new folder will be created with the provided name.}
}
\value{
Three csv tables of corresponding data in TSIS.data.
}
\description{
Save TSIS.data data into csv files
}
\examples{
TSIS.data.example(dir='data')
}
