% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSIS.data.R
\docType{data}
\name{TSIS.data}
\alias{TSIS.data}
\title{Example datasets of TSIS.data.}
\format{An object of data list}
\usage{
data(TSIS.data)
}
\description{
The TSIS package provides the example datasets "TSIS.data" with 30 genes and 140 isoforms, analyzed in 26 time points, each with 9 replicates. 
The isoform expression is in TPM (transcript per million) format. Other type of transcript quantifications, such as read counts, 
Percentage Splicing Ins (PSIs) can also be used in TSIS.
}
\examples{
names(TSIS.data)
##expression data
TSIS.data$data.exp[1:10,1:5]
##mapping data
TSIS.data$mapping[1:10,]
##subset of isoform names
TSIS.data$sub.isoforms[1:10]


}
\keyword{datasets}
