% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtRTD2.R
\docType{data}
\name{AtRTD2}
\alias{AtRTD2}
\title{Example datasets of AtRTD2.}
\format{An object of data list}
\usage{
data(AtRTD2)
}
\description{
The example datasets "AtRTD2" with 300 genes and 766 isoforms, analysed in 26 time points, each with 3 biological replicates and 3 technical replicates.
The experiments were designed to investigate the Arabidopsis gene expression response to cold. The isoform expression is in TPM (transcript per million) format.
For the experiments and data quantification details, please see the AtRTD2 paper (Zhang, et al.,2016).
}
\examples{
names(AtRTD2)
##expression data
AtRTD2$data.exp[1:10,1:5]
##mapping data
AtRTD2$mapping[1:10,]
##subset of isoform names
AtRTD2$sub.isoforms[1:10]
}
\references{
Zhang, R., et al. AtRTD2: A Reference Transcript Dataset for accurate quantification of alternative splicing and expression changes
in Arabidopsis thaliana RNA-seq data. bioRxiv 2016.
}
\keyword{datasets}

