% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.switch.R
\name{iso.switch}
\alias{iso.switch}
\title{Isoform switch analysis for time-series data}
\usage{
iso.switch(data.exp, mapping, t.start = 1, t.end = 26, nrep = 9,
  rank = F, min.t.points = 2, min.distance = 1, spline = F,
  spline.df = NULL, verbose = T)
}
\arguments{
\item{data.exp}{isoform expression data frame with row names of isoforms and column names of samples.}

\item{mapping}{gene-isoform mapping data frame with first column of genes and second column of isoforms.}

\item{t.start, t.end}{start and end time points of the time-series data. The time step is assumed to be 1.}

\item{nrep}{number of replicates for each time point.}

\item{rank}{logical, to use rank of isoform expression for each sample (TRUE) or not (FALSE).}

\item{min.t.points}{pre-filtering, if the number of time points in all intervals < \code{min.t.points}, skip this pair of isoforms.}

\item{min.distance}{pre-filtering, if the sample distances in the time courses (mean expression or splined value) for
intersection search all < \code{min.distance}, skip this pair of isoforms.}

\item{spline}{logical, to use spline method (TRUE) or mean expression (FALSE).}

\item{spline.df}{the degree of freedom used in spline method. See \code{\link{ns}} in \code{\link{splines}} for details.}

\item{verbose}{logical, to track the progressing of runing (TRUE) or not (FALSE).}
}
\value{
a data frame of scores. The column names:
\itemize{
\item{iso1,iso2: }{the isoforms.}
\item{iso1.mean.ratio, iso2.mean.ratio: }{the ratio of isoforms to their genes.}
\item{left.interval, left.interval: }{the left (before switch) and right (after switch) intervals of a switch point.}
\item{x.value, y.value: }{x and y coordinates of switch points.}
\item{left.prob, right.prob: }{the frequencies/probabilities that the samples of a isoform is greater or less than the other in left and right intervals, respectively. }
\item{left.dist, right.dist: }{the average sample distances in left and right intervals, respectively. }
\item{left.pval, right.pval: }{p-values of paired t-test for samples in left and right intervals, respectively. }
\item{left.t.points, right.t.points: }{number of time points in left and right intervals, respectively. }
\item{prob: }{Score1.}
\item{dist: }{Score2.}
\item{cor: }{Pearson correlation of two isoforms.}

}
}
\description{
This function is used to search and score isoform switch points in time-series isoform expression data.
}
\details{
The detailed steps:

\bold{Step 1: search for time course intersection points.}

The expression for a pair of isoforms iso1 and iso2 may experience a number isoform switch in the whole
time duration. Two methods have been included to search for these switch points where the isoforms reverse relative
expression profiles.
\itemize{
\item{\bold{Method 1:}}{ use average expression values across time points. Taking average values of the replicates
for time points in the input isoform expression data.}
\item{\bold{Method 2:}}{ use nature spline curves to fit the time-series data and find intersection points of the
fitted curves for each pair of isoforms. See details in \code{\link{ts.spline}}}
}

\bold{Step 2: score the isoform switches}

We defined 5 parameters to score the quality of isoform switch. The first two are the frequency of switch and the
sum of average distance before and after switch, used as Score 1 and Score 2 in iso-kTSP \url{https://bitbucket.org/regulatorygenomicsupf/iso-ktsp}
method for two condition comparisons (Sebestyen, et al., 2015).
To investigate the switches of two isoforms \eqn{iso_i} and \eqn{iso_j} in two conditions \eqn{c_1} and \eqn{c_2}, \bold{Score 1} is defined as
\deqn{S_1(iso_i,iso_j|c_1,c_2)=|p(iso_1>iso2|c_1)+p(iso_1<iso_2|c_2)-1|}
where \eqn{p(iso_1>iso2|c_1)} and \eqn{p(iso_1<iso_2|c_2)} are the frequencies/probabilities that the samples of one isoform
is greater or less than the other in corresponding conditions. \bold{Score 2} is defined as
\deqn{S_2(iso_i,iso_j|c_1,c_2)=|mean.dist(iso_i,iso_2|c_1)|+|mean.dist(ios_1,iso_2|c_2)|}
where \eqn{mean.dist(iso_i,iso_2|c_1)} and \eqn{mean.dist(ios_1,iso_2|c_2)} are the mean distances of samples in conditions \eqn{c_1} and \eqn{c_2}, respectively.

However, the time-series for a pair of isoforms may undergo a number of switch points in the time duration. To extend the
iso-kTSP to TSIS, the time duration is divided in to intervals with the intersection points determined in Step 1.
For each pair of consecutive intervals before and after switch, they can be assimilated as two conditions to
implement the calculation of Score 1 and Scoe 2.

The time-series isoform switches are more complex than the comparisons over two conditions. In addition to
Score 1 and Score 2 for each switch point, we defined other 3 parameters as metrics of switch qualities.

\itemize{
\item{\bold{Score 3:}}{ p-value of paired t-test for the two isoform sample differences within each interval.}
\item{\bold{Score 4:}}{ Time points number within each interval.}
\item{\bold{Score 5:}}{ Pearson correlation of two isoforms.}
}
Note: since each switch point has a left and right adjoined intervals before and after
switch, two p-values and two numbers of time points for the intervals are assigned to each
switch point, respectively.
}
\references{
1.	Sebestyen E, Zawisza M, Eyras E: Detection of recurrent alternative splicing switches in tumor samples reveals novel signatures of cancer.
Nucleic Acids Res 2015, 43(3):1345-1356.
}

