% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch.density.R
\name{switch.density}
\alias{switch.density}
\title{Density plot}
\usage{
switch.density(x, make.plotly = T, ...)
}
\arguments{
\item{x}{a numeric vector to plot th density.}

\item{make.plotly}{logical, to plot \code{\link{plotly}} format figures (TRUE)
or general plot (FALSE)?. See details in \code{\link{ggplotly}} in \code{\link{plotly}} R pacakge.}

\item{...}{additional parameters pass to \code{\link{density}}}
}
\value{
density plot in general format or \code{\link{plotly}} format if \code{make.plotly=T}
}
\description{
Density plot
}
\examples{
##random values
set.seed(100)
x<-rnorm(100,mean=0,sd=1)
##geneal format
switch.density(x,make.plotly =F)
##plotly format
switch.density(x,make.plotly =T)


}

